# Python: Use `-O3` level optimization