--- a/mingw-w64-python/PKGBUILD
+++ b/mingw-w64-python/PKGBUILD
@@ -355,7 +355,7 @@ build() {
     _extra_config+=("--enable-optimizations")
     # PGO should be done with -O3
     CFLAGS="${CFLAGS/-O2/-O3}"
-    CFLAGS+=$([[ ${MINGW_PACKAGE_PREFIX} == *-clang* ]] || echo " -ffat-lto-objects")
+    # CFLAGS+=$([[ ${MINGW_PACKAGE_PREFIX} == *-clang* ]] || echo " -ffat-lto-objects")
     # FIXME: https://github.com/msys2-contrib/cpython-mingw/issues/10
     # _extra_config+=("--with-lto")
   else