--- a/mingw-w64-python/PKGBUILD
+++ b/mingw-w64-python/PKGBUILD
@@ -353,8 +353,10 @@ build() {
   declare -a _extra_config
   if check_option "debug" "n"; then
     _extra_config+=("--enable-optimizations")
+
     # PGO should be done with -O3
-    CFLAGS="${CFLAGS/-O2/-O3}"
+    # CFLAGS="${CFLAGS/-O2/-O3}"
+    CFLAGS+=" -O3"
     # CFLAGS+=$([[ ${MINGW_PACKAGE_PREFIX} == *-clang* ]] || echo " -ffat-lto-objects")
     # FIXME: https://github.com/msys2-contrib/cpython-mingw/issues/10
     # _extra_config+=("--with-lto")