@naveen521kk 2022-06-14
I think this PR should wait until the Python 3.10 update rebuild is done(see https://packages.msys2.org/queue). Also, lazka is unavailable atm and will soon review this PR.